% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_ncf.R
\name{deep_ncf}
\alias{deep_ncf}
\title{Deep Learning Recommendation: Neural Collaborative Filtering}
\usage{
deep_ncf(data, num_layer = seq(1, 5, 1), max_units = 10, start_unit = 3,
  max_lr = 0.2, min_lr = 0.001, iteration_per_layer = 5,
  min_mf_output_dim = 2, max_mf_output_dim = 10, num_epoch = 2,
  top = 100)
}
\arguments{
\item{data}{Transaction dataset. Format must be order_id (could be dummy), user_id, and product_id}

\item{num_layer}{a vector of integers indicating the number of hidden layers to test. Default to seq(1,5,1)}

\item{max_units}{the maximum number of hidden units in a layer. Default to 10}

\item{start_unit}{the minimum number of hiddent units in a layer. Default to 3}

\item{max_lr}{maximum learning rate in a run. Default to 0.2}

\item{min_lr}{minimum learning rate in a run. Default to 0.001}

\item{iteration_per_layer}{Number of parameter randomizations for a given number of hidden layers. More iterations will explore a larger parameter space}

\item{min_mf_output_dim}{Min number of latent factors to represent users and items}

\item{max_mf_output_dim}{Max number of latent factors to represent users and items}

\item{num_epoch}{number of epoches to go through during training}

\item{top}{Number of top products and customers to include. Can be NULL (e.g. will include everything)}
}
\value{
returns a list object with two values: 
train_performance: A table with parameters and model performance metrics
best_model: a keras_model object with the optimal parameters
}
\description{
Deep Learning Recommendation: Neural Collaborative Filtering
}
