% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_classification_single.R
\name{dl_classification_single}
\alias{dl_classification_single}
\title{Deep Learning Classification with Known Network Structure}
\usage{
dl_classification_single(x, y, complexity, dropout, lr, num_epoch = 5,
  num_patience = 3)
}
\arguments{
\item{x}{training feature matrix}

\item{y}{target matrix}

\item{complexity}{a vector indicating numbers of hidden units in each layer, e.g. c(3,6,7) means 3 layers with 3, 6, 7 units in each layer}

\item{dropout}{a vector indicating the dropout rate in each layer, e.g c(0.1,0.2,0.3)}

\item{lr}{learning rate for the optimizer}

\item{num_epoch}{number of epoches to go through during training}

\item{num_patience}{number of patience in early stopping criteria}
}
\value{
returns a list object with three values:
model: keras model contructed. A keras_model object
loss: a vector containing loss value in each epoch
accuracy: a vector containing accuracy value in each epoch
}
\description{
Deep Learning Classification with Known Network Structure
}
