% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncf_data_transform.R
\name{ncf_data_transform}
\alias{ncf_data_transform}
\title{NCF Utility Function: Transform transaction data to a data with 0-1 labels}
\usage{
ncf_data_transform(data, top = 1000)
}
\arguments{
\item{data}{Transaction dataset. Format must be order_id (could be dummy), user_id, and product_id}

\item{top}{Number of top products and customers to include. Can be NULL (e.g. will include everything)}
}
\value{
returns a dataset with user_id, product_id, label (0,1 indicating if past purchase exists)
}
\description{
NCF Utility Function: Transform transaction data to a data with 0-1 labels
}
