% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncf_get_training_data.R
\name{ncf_get_training_data}
\alias{ncf_get_training_data}
\title{NCF Utility Function: Subset negative cases in a dataset}
\usage{
ncf_get_training_data(data, negative_ratio)
}
\arguments{
\item{data}{a data frame with positive and negative label, must have a column named label}

\item{negative_ratio}{ratio of desired number of negative cases vs positive cases}
}
\value{
returns a dataset with the desired negative-to-positive ratio
}
\description{
NCF Utility Function: Subset negative cases in a dataset
}
