% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncf_single.R
\name{ncf_single}
\alias{ncf_single}
\title{Neural Collaborative Filtering with a single configuration}
\usage{
ncf_single(data, mf_output_dim, mlp_complexity, lr = 0.001, epoch = 2,
  top = 100)
}
\arguments{
\item{data}{Transaction dataset. Format must be order_id (could be dummy), user_id, and product_id}

\item{mf_output_dim}{Number of latent vectors to represent users and items}

\item{mlp_complexity}{Number of hidden layers in the MLP layer}

\item{top}{Number of top products and customers to include. Can be NULL (e.g. will include everything)}
}
\value{
returns a list object with three values: 
model: keras model contructed. A keras_model object
loss: a vector containing loss value in each epoch
}
\description{
Neural Collaborative Filtering with a single configuration
}
