% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_numeric_col_trans.R
\name{non_numeric_col_trans}
\alias{non_numeric_col_trans}
\title{Transform Non-numeric Columns to Hot Encoding}
\usage{
non_numeric_col_trans(data, id.var, target.var = NULL)
}
\arguments{
\item{data}{data frame}

\item{id.var}{name of the column to indicate unique row}

\item{target.var}{name of the target column. Default to NULL}
}
\value{
a data frame with transformed columns
}
\description{
Transform Non-numeric Columns to Hot Encoding
}
