% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_train.R
\name{write_train}
\alias{write_train}
\title{Write train.R file for cloudml_train}
\usage{
write_train(num_layer, num_epoch, num_patience, target_type)
}
\arguments{
\item{num_layer}{Number of hidden layer_dense}

\item{num_epoch}{number of epoches to go through during training}

\item{num_patience}{number of patience in early stopping criteria}

\item{target_type}{either classification or regression}
}
\description{
Write train.R file for cloudml_train
}
