% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yml.R
\name{write_yml}
\alias{write_yml}
\title{Write config.yml file for cloudml_train}
\usage{
write_yml(num_layer, max_units, start_unit, max_dropout, min_dropout, max_lr,
  min_lr)
}
\arguments{
\item{num_layer}{Number of hidden layer_dense}

\item{max_units}{the maximum number of hidden units in a layer. Default to an optimized value based on data}

\item{start_unit}{the minimum number of hiddent units in a layer. Default to 5}

\item{max_dropout}{A number between 0 and 1 indicating the maximum dropoff rate in a layer. Default to 0.2}

\item{min_dropout}{A number between 0 and 1 indicating the minimum dropoff rate in a layer. Default to 0}

\item{max_lr}{maximum learning rate in a run. Default to 0.2}

\item{min_lr}{minimum learning rate in a run. Default to 0.001}
}
\description{
Write config.yml file for cloudml_train
}
