% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdl_single.R
\name{cdl_single}
\alias{cdl_single}
\title{Collaborative Deep Learning with a single network configuration}
\usage{
cdl_single(items, rating)
}
\arguments{
\item{items}{item matrix with numeric item attributes}

\item{rating}{user-item rating matrix 
return return a user-item ratings with scores for previously unseen items for each user}
}
\description{
Collaborative Deep Learning with a single network configuration
}
