% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_collect.R
\name{google_collect}
\alias{google_collect}
\title{Collect model output from Google Cloud}
\usage{
google_collect(model_id)
}
\arguments{
\item{model_id}{a string output from the previous function indicating the model_id}
}
\description{
Collect model output from Google Cloud
}
